move test_gp
