+++ test/unit/optimizer/gp_operators/test_crossover.py
@@ -5,7 +5,7 @@ from fedot.core.optimisers.gp_comp.individual import Individual
 from fedot.core.optimisers.gp_comp.operators.crossover import CrossoverTypesEnum, Crossover
 from fedot.core.optimisers.gp_comp.pipeline_composer_requirements import PipelineComposerRequirements
 from fedot.core.pipelines.pipeline_graph_generation_params import get_pipeline_generation_params
-from test.unit.optimizer.test_gp_operators import generate_pipeline_with_single_node
+from test.unit.optimizer.gp_operators.test_gp_operators import generate_pipeline_with_single_node
 from test.unit.pipelines.test_node_cache import pipeline_first, pipeline_second
